+++ plone/app/event/tests/test_event_listing.py
@@ -7,28 +7,31 @@ from plone.app.event.tests.base_setup import patched_now as PN
 import mock
 
 
-class TestEventsListingDX(AbstractSampleDataEvents):
+class TestEventsListingPortal(AbstractSampleDataEvents):
     layer = PAEventDX_INTEGRATION_TESTING
+
+    def _listing_view(self, name='@@event_listing'):
+        return self.portal.restrictedTraverse(name)
 
     @mock.patch('plone.app.event.browser.event_listing.localized_now', new=PN)
     @mock.patch('plone.app.event.base.localized_now', new=PN)
     def test_get_events_future(self):
         # Default mode is to show all events from now on.
-        view = self.portal.restrictedTraverse('@@event_listing')
+        view = self._listing_view()
         self.assertEqual(len(view._get_events()), 5)
 
     @mock.patch('plone.app.event.browser.event_listing.localized_now', new=PN)
     @mock.patch('plone.app.event.base.localized_now', new=PN)
     def test_get_events_past(self):
         self.request.form.update({'mode': 'past'})
-        view = self.portal.restrictedTraverse('@@event_listing')
+        view = self._listing_view()
         self.assertEqual(len(view._get_events()), 5)
 
     @mock.patch('plone.app.event.browser.event_listing.localized_now', new=PN)
     @mock.patch('plone.app.event.base.localized_now', new=PN)
     def test_get_events_all(self):
         self.request.form.update({'mode': 'all'})
-        view = self.portal.restrictedTraverse('@@event_listing')
+        view = self._listing_view()
         self.assertEqual(len(view._get_events()), 8)
 
     @mock.patch('plone.app.event.browser.event_listing.localized_now', new=PN)
@@ -36,7 +39,7 @@ class TestEventsListingDX(AbstractSampleDataEvents):
     def test_get_events_today(self):
         today = localized_today().isoformat()
         self.request.form.update({'mode': 'day', 'date': today})
-        view = self.portal.restrictedTraverse('@@event_listing')
+        view = self._listing_view()
         self.assertEqual(len(view._get_events()), 2)
 
     @mock.patch('plone.app.event.browser.event_listing.localized_now', new=PN)
@@ -44,9 +47,72 @@ class TestEventsListingDX(AbstractSampleDataEvents):
     def test_events_listing_ical(self):
         # Default mode is to show all events from now on.
         headers, output, request = make_fake_response(self.request)
-        view = self.portal.restrictedTraverse('@@event_listing_ical')
+        view = self._listing_view(name='@@event_listing_ical')
         view()
         self.assertEqual(len(headers), 2)
         self.assertEqual(headers['Content-Type'], 'text/calendar')
         icalstr = ''.join(output)
         self.assertTrue('Long Event' in icalstr)
+
+
+class TestEventsListingCollection(TestEventsListingPortal):
+
+    def _listing_view(self, name='@@event_listing'):
+        return self.portal.collection.restrictedTraverse(name)
+
+    def test_date_filtering(self):
+        """Test if date filters are available on Collections without start or
+        end search criterias.
+        """
+        # plone.app.contenttypes ICollection type
+        self.portal.invokeFactory(
+            'Collection', 'col_without_date_criterion', title=u'Col')
+        collection = self.portal.col_without_date_criterion
+        collection.query = [
+            {'i': 'portal_type',
+             'o': 'plone.app.querystring.operation.selection.is',
+             'v': ['Event', 'plone.app.event.dx.event']
+             },
+        ]
+        view = self.portal.col_without_date_criterion.restrictedTraverse(
+            '@@event_listing'
+        )
+        out = view()
+        for _class in [
+            'mode_future',
+            'mode_past',
+            'mode_month',
+            'mode_week',
+            'mode_day',
+        ]:
+            self.assertTrue(_class in out)
+
+    def test_no_date_filtering(self):
+        """Test if date filters are not available on Collections with start or
+        end search criterias.
+        """
+        # plone.app.contenttypes ICollection type
+        self.portal.invokeFactory(
+            'Collection', 'col_with_date_criterion', title=u'Col')
+        collection = self.portal.col_with_date_criterion
+        collection.query = [
+            {'i': 'portal_type',
+             'o': 'plone.app.querystring.operation.selection.is',
+             'v': ['Event', 'plone.app.event.dx.event']
+             },
+            {'i': 'start',
+             'o': 'plone.app.querystring.operation.date.afterToday',
+             'v': ''}
+        ]
+        view = self.portal.col_with_date_criterion.restrictedTraverse(
+            '@@event_listing'
+        )
+        out = view()
+        for _class in [
+            'mode_future',
+            'mode_past',
+            'mode_month',
+            'mode_week',
+            'mode_day',
+        ]:
+            self.assertTrue(_class not in out)